Awakening Fixes
Added a longer Delay to the Harvester Assignment because it caused crashes.
Gave the Temple Prime and Cyborg Factories back to the Nod Player, so any spawned Cyborgs get equally assigned to all Players.
Also changed the Power Use of the Temple Prime to 0, so it wont power down.

And I added Lobby Options to increase the amount of waves and airstrikes due to popular demand. :-)